Let VectorWindow be copy- and resizeable
Though this is not the intended use of a VectorWindow, certain algorithms
in other module make frequently use of the possibility to copy (and sometimes even resize)
a block. While I dot not encourage this, reimplementing everything is probably not worth
the effort.
If you copy-construct or resize a VectorWindow, the resulting object will be a window to
memory it now manages _itself_. Memory will be de-allocated if the object goes out of scope
through the unique_ptr mechanism.